Lagre agent likt i endringshistorikk som i totrinnskontroll
