+++ common/domain/src/main/kotlin/no/nav/mulighetsrommet/model/Agent.kt
@@ -12,3 +12,20 @@ data object Arena : Agent
 
 @Serializable
 data object Arrangor : Agent
+
+fun String.toAgent(): Agent = when (this) {
+    "Tiltaksadministrasjon" -> Tiltaksadministrasjon
+    "Arena" -> Arena
+    "Arrangor" -> Arrangor
+    else -> NavIdent(this)
+}
+
+/* Format used when stored in text column in db */
+fun Agent.textRepr(): String = when (this) {
+    Arena,
+    Arrangor,
+    Tiltaksadministrasjon,
+    ->
+        this.toString()
+    is NavIdent -> this.value
+}
