+++ mulighetsrommet-api/src/main/resources/db/migration/V247__endringshistorikk_agent.sql
@@ -1,0 +1,2 @@
+update endringshistorikk set user_id = 'Arrangor' where user_id = 'Arrangør';
+update endringshistorikk set user_id = 'Tiltaksadministrasjon' where user_id = 'System';
